requirements.txt
python-vlc
gtts